tomoor_size_detect
This code is for building a Convolutional Neural Network (CNN) model for medical image classification. It loads images from a directory structure where each patient's data is organized into subdirectories based on different tumor sizes. The images are preprocessed, normalized, and then used to train the CNN model.

The CNN model architecture consists of convolutional layers followed by max-pooling layers for feature extraction, and then dense layers for classification. It's compiled with the Adam optimizer and categorical cross-entropy loss. The model is trained using the training data and evaluated using the test data. 

After training, the code generates a confusion matrix to visualize the model's performance on the test data. Additionally, it allows for making predictions on new images by loading, preprocessing, and passing them through the trained model. The predicted class index is then mapped to a corresponding folder structure.